Prépare au test en CI

+++ provision.sh
@@ -1,12 +1,24 @@
 #!/bin/bash
 
-set +ev
+set -ev
+
+# Info
+pwd
+ls -al
+
+apt-get update
+apt-get install --assume-yes openssh-client python3-pip rsync vim
+mkdir --parents /opt/mes-aides
+
+if [ test -e /vagrant/ ];
+then
+  rsync -r /vagrant/ /opt/mes-aides/ops
+else
+  rsync -r . /opt/mes-aides/ops
+fi
 
 cd
 ssh-keygen -t rsa -q -f "$HOME/.ssh/id_rsa" -m PEM -N "" -C "contact@mes-aides.gouv.fr"
-apt-get install --assume-yes python3-pip vim
-mkdir --parents /opt/mes-aides
-rsync -r /vagrant/ /opt/mes-aides/ops
 cd /opt/mes-aides/ops
 pip3 install --requirement requirements.txt
 cat "$HOME/.ssh/id_rsa.pub" >>  "$HOME/.ssh/authorized_keys"
